Move restapi to examples/applications
- Removed other example applications as per rt's directions
- Changed @routes decorator syntax to adding routes at top of file
- Replace over-use of 'isinstance' with try/except/raise Fault
