+++ electrumsv/storage.py
@@ -1243,7 +1243,6 @@ class WalletStorage:
 
     def __init__(self, path: str, manual_upgrades: bool=False,
             storage_kind: StorageKind=StorageKind.UNKNOWN) -> None:
-
         logger.debug("wallet path '%s'", path)
         dirname = os.path.dirname(path)
         if not os.path.exists(dirname):
