--- examples/applications/esv_fileupload/__init__.py
@@ -1,3 +1,0 @@
-from .app_state import LocalAppStateProxy
-
-# The AppStateProxy subclass is identifed by ElectrumSV as the entrypoint.
