--- examples/applications/esv_merchant_server/__init__.py
@@ -1,3 +1,0 @@
-from .app_state import MerchantAppStateProxy
-
-# The AppStateProxy subclass is identifed by ElectrumSV as the entrypoint.
