--- examples/applications/esv_merchant_server/app_state.py
@@ -1,13 +1,0 @@
-from electrumsv.app_state import AppStateProxy
-
-from .app import MerchantApplication
-
-
-class MerchantAppStateProxy(AppStateProxy):
-    def __init__(self, *args) -> None:
-        super().__init__(*args)
-
-        self.app = MerchantApplication()
-
-    def has_app(self):
-        return True
