--- examples/applications/esv_merchant_server/rpc.py
@@ -1,8 +1,0 @@
-from electrumsv.logs import logs
-
-class LocalRPCFunctions:
-    def __init__(self) -> None:
-        self._logger = logs.get_logger("local-rpc")
-
-    def test123(self):
-        return 123
