+++ examples/applications/restapi/__init__.py
@@ -1,0 +1,3 @@
+from .app_state import LocalAppStateProxy
+
+# The AppStateProxy subclass is identifed by ElectrumSV as the entrypoint.
