+++ examples/applications/restapi/app_state.py
@@ -1,13 +1,14 @@
+from electrumsv.async_ import ASync
 from electrumsv.app_state import AppStateProxy
 
-from .app import FileUploadApplication
+from .app import RESTAPIApplication
 
 
 class LocalAppStateProxy(AppStateProxy):
     def __init__(self, *args) -> None:
         super().__init__(*args)
-
-        self.app = FileUploadApplication()
+        self.async_ = ASync()
+        self.app = RESTAPIApplication()
 
     def has_app(self):
         return True
